Shift focus to subscribe/sign popup notification
If a user is not logged in and visits a free article, a 'Please
subscribe or sign in...' message appears over the header when they
attempt to add the article to myFT.

Previously, when the message appeared the focus would remain present on
the 'Add to myFT' button and would not shift to the message. This is a
problem for keyboard users.

This PR fixes this by shifting the focus to the message so it's
accessible to keyboard users. It's done by passing in the new
n-noticifation focusSelector and returnFocusSelector properties.

 🐿 v2.12.3