Concept lists are supposed to have black links
Concept lists were using a load of `n-content-tag` classes which, as a
result of the origami change, were making links claret. It turns out we
only really needed the bold and some margin-right from those tags. So
rather than use them, directly style the links as we want them. 🐿 v2.12.5