Implement create list and save article modal variant
Adds a data property data-myft-ui-variant=createListAndSaveArticleVariant to the save for later button that will determine if the test is active based on the value of the professorLists flag.
In the myft.user.saved.content.add event listener, checks for the data property on the save for later button and displays the new modal variant if the user has no lists.
Implement the modal variant using o-overlay and implement methods to: position the modal next to the SaveForLater component, reposition the modal on resize or scroll (mobile only) handle the interactions to create a new list and add/remove the article to lists.