Fix scroll listener issue on saveArticleToList variant
The realignOverlay method is called on a scroll listener to position the overlay above or below the share nav. This method was using the overlay position to determine which half of the screen has more space. This approach was causing a flickering issue on the overlay. This commit changes it to use the shareNav component position instead.