Add Public toggle to new list form on manageArticleLists modal
Adds the markup for the "Public" toggle to the FormElement
Captures the Public toggle state and sends it to the createList method. The createList method will use it to display a fake door  message
Changes the markup and style of the FormElement to add the design changes required for this feature

This toggle is part of a smoke test. For now, it doesn't change anything on the created list, but it'll help us measure interest in this feature