Use value of myftListPublicPrivateToggle to determine if the public/private toggle should show on manageArticleLists component
Adds the attribute data-myft-ui-save-new-config to the Save button. We'll use it to add configuration props to the manageArticleLists component
Adds the value showToggleProp to that attribute if the myftListPublicPrivateToggle flag is on
Passes a config object to the manageArticleLists. Passes the name and the showToggleProp properties through it.
Conditionally shows the Public toggle on the FormElement if the showToggleProp is true